fix: invert pwm command instead of effort cmd
Inverting the effort command would result in one of the motors not
achieving acting in a full thrust pwm command. That happens because the
effort to pwm table isnt symmetric (effort goes from 16.44 to -10.21).

This was not caught by the existing unit tests because they were using
wrong values in the "saturated" tests